infovoid/requirements/dev.txt
-r tests.txt

django-debug-toolbar==1.11 
